refactor: Improve error handling and session management in user registration and update controllers

+++ src/main/java/br/com/lux/services/user/login/LoginService.java
@@ -42,41 +42,63 @@ public class LoginService implements UserService
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public List<User> findAllUsers()
     {
-        return userRepository.findAll();
+        try
+        {
+            return userRepository.findAll();
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao buscar todos os usuários! " + e.getMessage());
+        }
     }
 
     @Override
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public Optional<User> authenticate(String email, String password)
     {
-        Optional<User> optionalUser = userRepository.findByEmail(email);
+        try
+        {
+            Optional<User> optionalUser = userRepository.findByEmail(email);
 
-        if(optionalUser.isPresent() && PasswordUtils.checkPassword(password, optionalUser.get().getPassword()))
-            return Optional.of(optionalUser.get());
+            if(optionalUser.isPresent() && PasswordUtils.checkPassword(password, optionalUser.get().getPassword()))
+                return Optional.of(optionalUser.get());
 
-        return Optional.empty();
+            return Optional.empty();
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao autenticar usuário! " + e.getMessage());
+        }
     }
 
     @Override
     @Transactional
     public Optional<User> createUser(User user)
     {
-        Optional<User> existingUser = userRepository.findByEmail(user.getEmail());
-
-        if (existingUser.isPresent())
+        try
         {
-            return Optional.empty();
+            Optional<User> existingUser = userRepository.findByEmail(user.getEmail());
+
+            if (existingUser.isPresent())
+            {
+                return Optional.empty();
+            }
+            else
+            {
+                String urlAvatar = gravatarService.getGravatarUrl(user.getEmail().toLowerCase());
+
+                user.setUrlavatar(urlAvatar);
+                user.setTipo(UserType.CLIENTE);
+                user.setPassword(PasswordUtils.encryptPassword(user.getPassword()));
+
+                userRepository.save(user);
+
+                return Optional.of(user);
+            }
         }
-        else
+        catch (Exception e)
         {
-            String urlAvatar = gravatarService.getGravatarUrl(user.getEmail().toLowerCase());
-            user.setUrlavatar(urlAvatar);
-            user.setTipo(UserType.CLIENTE);
-            user.setPassword(PasswordUtils.encryptPassword(user.getPassword()));
-
-            userRepository.save(user);
-
-            return Optional.of(user);
+            throw new ServiceException("Erro ao criar usuário! " + e.getMessage());
         }
     }
 
@@ -84,13 +106,19 @@ public class LoginService implements UserService
     @Transactional
     public void changeClientId(Integer id, Client client)
     {
-        Optional<User> optionalUser = userRepository.findById(id);
-
-        if(optionalUser.isPresent())
+        try
         {
-            User user = optionalUser.get();
-            user.setCliente(client);
-            userRepository.save(user);
+            Optional<User> optionalUser = userRepository.findById(id);
+            if(optionalUser.isPresent())
+            {
+                User user = optionalUser.get();
+                user.setCliente(client);
+                userRepository.save(user);
+            }
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao alterar id do cliente! " + e.getMessage());
         }
     }
 
@@ -98,14 +126,34 @@ public class LoginService implements UserService
     @Transactional
     public void deleteById(Integer id)
     {
-        userRepository.deleteById(id);
+       try
+       {
+            if (id == null || id == 1)
+                throw new ServiceException("Id inválido!");
+
+            userRepository.deleteById(id);
+       }
+       catch (ServiceException e)
+       {
+            throw new ServiceException("Erro ao deletar usuário por id! " + e.getMessage());
+       }
     }
 
     @Override
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public User findById(Integer id)
     {
-        return userRepository.findById(id).orElse(null);
+        try
+        {
+            if (id == null)
+                throw new ServiceException("Id inválido!");
+
+            return userRepository.findById(id).orElse(null);
+        }
+        catch (ServiceException e)
+        {
+            throw new ServiceException("Erro ao buscar usuário por id! " + e.getMessage());
+        }
     }
 
     @Override
@@ -142,15 +190,22 @@ public class LoginService implements UserService
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public String validarEmail(String email)
     {
-        if(email == null || email.isBlank())
-            return "Email inválido!";
+        try
+        {
+            if(email == null || email.isBlank())
+                return "Email inválido!";
 
-        Optional<User> optionalUser = userRepository.findByEmail(email);
+            Optional<User> optionalUser = userRepository.findByEmail(email);
 
-        if(optionalUser.isEmpty())
-            return "Email não cadastrado!";
+            if(optionalUser.isEmpty())
+                return "Email não cadastrado!";
 
-        return "null";
+            return "null";
+        }
+        catch (DataIntegrityViolationException e)
+        {
+            throw new ServiceException("Erro ao validar email! " + e.getMessage());
+        }
     }
 
     @Override
@@ -171,7 +226,7 @@ public class LoginService implements UserService
         }
         catch (Exception e)
         {
-            e.printStackTrace();
+            throw new ServiceException("Erro ao enviar código de verificação! " + e.getMessage());
         }
     }
 
@@ -179,25 +234,43 @@ public class LoginService implements UserService
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public boolean validarCodigoVerificacao(String email, String codigo, HttpSession session)
     {
-        String codigoSalvo = (String) session.getAttribute(email + "_verification_code_");
-
-        if (codigo.equals(codigoSalvo))
+        try
         {
-            session.removeAttribute(email + "_verification_code_");
-            return true;
-        }
+            String codigoSalvo = (String) session.getAttribute(email + "_verification_code_");
 
-        return false;
+            if (codigo.equals(codigoSalvo))
+            {
+                session.removeAttribute(email + "_verification_code_");
+                return true;
+            }
+
+            return false;
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao validar código de verificação! " + e.getMessage());
+        }
     }
 
     @Override
     @Transactional
     public void redefinirSenha(String email, String novaSenha)
     {
-        User user = userRepository.findByEmail(email).orElseThrow();
+        try
+        {
+            Optional<User> optionalUser = userRepository.findByEmail(email);
 
-        user.setPassword(PasswordUtils.encryptPassword(novaSenha));
+            if (userRepository.findByEmail(email).isEmpty())
+                throw new ServiceException("Email não cadastrado!");
 
-        userRepository.save(user);
+            User user = optionalUser.get();
+            user.setPassword(PasswordUtils.encryptPassword(novaSenha));
+
+            userRepository.save(user);
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao redefinir senha! " + e.getMessage());
+        }
     }
 }
